Make doubleTap and keyboard transitions exlusive to MapController
